+++ src/Propulsion.EventStore/EventStoreSink.fs
@@ -71,7 +71,7 @@ module Internal =
         let toStreams, oStreams = HashSet(), HashSet()
         let mutable okEvents, okBytes, exnEvents, exnBytes = 0, 0L, 0, 0L
 
-        override __.DumpStats() =
+        override _.DumpStats() =
             let results = !resultOk + !resultDup + !resultPartialDup + !resultPrefix
             log.Information("Completed {mb:n0}MB {completed:n0}r {streams:n0}s {events:n0}e ({ok:n0} ok {dup:n0} redundant {partial:n0} partial {prefix:n0} waiting)",
                 mb okBytes, results, okStreams.Count, okEvents, !resultOk, !resultDup, !resultPartialDup, !resultPrefix)
@@ -87,7 +87,7 @@ module Internal =
                 badCats.Clear(); resultExnOther := 0; oStreams.Clear()
             Equinox.EventStore.Log.InternalMetrics.dump log
 
-        override __.Handle message =
+        override this.Handle message =
             let inline adds x (set : HashSet<_>) = set.Add x |> ignore
             let inline bads streamName (set : HashSet<_>) = badCats.Ingest(StreamName.categorize streamName); adds streamName set
             base.Handle message
@@ -102,7 +102,7 @@ module Internal =
                 | Writer.Result.Duplicate _ -> incr resultDup
                 | Writer.Result.PartialDuplicate _ -> incr resultPartialDup
                 | Writer.Result.PrefixMissing _ -> incr resultPrefix
-                __.HandleOk res
+                this.HandleOk res
             | Scheduling.InternalMessage.Result (_duration, (stream, Choice2Of2 ((es, bs), exn))) ->
                 adds stream failStreams
                 exnEvents <- exnEvents + es
@@ -110,11 +110,11 @@ module Internal =
                 match Writer.classify exn with
                 | ResultKind.TimedOut -> adds stream toStreams; incr timedOut
                 | ResultKind.Other -> bads stream oStreams; incr resultExnOther
-                __.HandleExn exn
-        abstract member HandleOk : outcome : 'Outcome -> unit
-        default __.HandleOk(_ : 'Outcome) : unit = ()
-        abstract member HandleExn : exn : exn -> unit
-        default __.HandleExn(_ : exn) : unit = ()
+                this.HandleExn(log.ForContext("stream", stream).ForContext("events", es), exn)
+        abstract member HandleOk : Result -> unit
+        default _.HandleOk _ : unit = ()
+        abstract member HandleExn : log : ILogger * exn : exn -> unit
+        default _.HandleExn(_, _) : unit = ()
 
     type EventStoreSchedulingEngine =
         static member Create(log : ILogger, storeLog, connections : _ [], itemDispatcher, stats : Stats, dumpStreams, ?maxBatches, ?idleDelay)