+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Added
 ### Changed
+
+- `StreamsConsumerStats`: Standardized `HandleExn` signature and context [#110](https://github.com/jet/propulsion/pull/110)
+
 ### Removed
 ### Fixed